% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_helpers.R
\name{get_cycling_labels}
\alias{get_cycling_labels}
\title{Create labels for cycling levels}
\usage{
get_cycling_labels(hoods)
}
\arguments{
\item{hoods}{sf object containing neighborhood level cycling simulated data}
}
\value{
vector of character labels
}
\description{
Create labels for cycling levels
}
\details{
Creates quantile cycling labels for dummy cycling data
}
