% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc-datasets.R
\docType{data}
\name{hoods}
\alias{hoods}
\title{Neighborhood Data}
\format{
A \code{\link[sf]{sf}} object created from data at www.wprdc.org.
Downloaded 11/2021 ; previous modified datestamp 3/9/2021  9:02 PM
}
\usage{
hoods
}
\description{
Neighborhood Data
}
\seealso{
https://data.wprdc.org/dataset/neighborhoods2
}
\keyword{datasets}
