% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc-datasets.R
\docType{data}
\name{streets}
\alias{streets}
\title{Street Data}
\format{
A \code{\link[sf]{sf}} object created from data at www.wprdc.org. 
 Downloaded 11/2021 ; previous modified datestamp March 9, 2021, 5:00 PM
}
\usage{
streets
}
\description{
Street Data
}
\seealso{
https://data.wprdc.org/dataset/allegheny-county-owned-roads-centerlines1
}
\keyword{datasets}
